+++ src/App.js
@@ -6,12 +6,29 @@ import {
   StackNavigator
 } from "react-navigation";
 
-import SplashPage from "./containers/SplashPage";
-import HomePage from "./containers/HomePage";
-import SearchPage from "./containers/SearchPage";
-import LoginPage from "./containers/LoginPage";
-import SignUpPage from "./containers/SignUpPage";
-import MeetingPage from "./containers/MeetingPage";
+// shared routes
+import SplashPage from "./containers/shared/SplashPage";
+import HomePage from "./containers/shared/HomePage";
+import MeetingPage from "./containers/shared/MeetingPage";
+
+//anonymous routes
+import SearchPage from "./containers/anonymous/SearchPage";
+import LoginPage from "./containers/anonymous/LoginPage";
+import SignUpPage from "./containers/anonymous/SignUpPage";
+
+//signedinroutes
+
+import SettingsPage from "./containers/signedin/SettingsPage";
+import UserPage from "./containers/signedin/SettingsPage/UserPage";
+import CalendarPage from "./containers/signedin/SettingsPage/CalendarPage";
+import NotificationPage from "./containers/signedin/SettingsPage/NotificationPage";
+import InformationPage from "./containers/signedin/InformationPage";
+import InformationDetailsPage from "./containers/signedin/InformationPage/InformationDetailsPage";
+import SchedulePage from "./containers/signedin/SchedulePage";
+import ScheduleDetailsPage from "./containers/signedin//SchedulePage/ScheduleDetailsPage";
+import InboxPage from "./containers/signedin/InboxPage";
+import InboxDetailsPage from "./containers/signedin/InboxPage/InboxDetailsPage";
+import CheckInPage from "./containers/signedin/CheckInPage";
 import { SideMenu } from "../src/components";
 
 const SCREEN_WIDTH = Dimensions.get("window").width;
@@ -22,35 +39,26 @@ class App extends Component {
   }
 }
 
-// const AppDrawerNavigator = createDrawerNavigator(
-//   {
-//     HomePage: { screen: HomePage },
-//     SideMenu: { screen: SideMenu }
-//   },
-//   {
-//     contentComponent: SideMenu
-//   }
-// );
-
-// const AppStack = StackNavigator(
-//   {
-//     SplashPage: props => <SplashPage {...props} />,
-//     HomePage: props => <HomePage {...props} />,
-//     SearchPage: props => <SearchPage {...props} />,
-//     LoginPage: props => <LoginPage {...props} />,
-//     SignUpPage: props => <SignUpPage {...props} />,
-//     MeetingPage: props => <MeetingPage {...props} />,
-//     SideMenu: props => <SideMenu {...props} />,
-//     Drawer: { screen: AppDrawerNavigator }
-//   },
-//   {
-//     initialRouteName: "SplashPage",
-//     headerMode: "none",
-//     navigationOptions: {
-//       headerVisible: false
-//     }
-//   }
-// );
+const SignedInStack = createStackNavigator(
+  {
+    SplashPage: props => <SplashPage {...props} />,
+    HomePage: props => <HomePage {...props} />,
+    MeetingPage: props => <MeetingPage {...props} />,
+    SettingsPage: props => <SettingPage {...props} />,
+    // insert user, calendar and notification if in use
+    NotificationPage: props => <SettingPage {...props} />,
+    InformationPage: props => <InformationPage {...props} />,
+    SchedulePage: props => <SchedulePage {...props} />,
+    InboxPage: props => <InboxPage {...props} />,
+    CheckInPage: props => <CheckInPage {...props} />
+  },
+  {
+    headerMode: "none",
+    navigationOptions: {
+      headerVisible: false
+    }
+  }
+);
 
 const AnonymousStack = createStackNavigator(
   {
@@ -72,7 +80,8 @@ const AnonymousStack = createStackNavigator(
 
 const RootStack = createDrawerNavigator(
   {
-    AnonymousStack: { screen: AnonymousStack }
+    AnonymousStack: { screen: AnonymousStack },
+    SignedInStack: { screen: SignedInStack }
   },
   {
     drawerWidth: SCREEN_WIDTH * 0.8,
