+++ src/components/MenuButton/index.js
@@ -2,9 +2,13 @@ import React from "react";
 import { View, TouchableOpacity, Image } from "react-native";
 import ComponentStyle from "./styles";
 
-const MenuButton = ({ onPress, image }) => {
+const MenuButton = ({ onPress, image, status }) => {
   return (
-    <View style={ComponentStyle.container}>
+    <View
+      style={
+        status === "SI" ? ComponentStyle.containerSI : ComponentStyle.containerA
+      }
+    >
       <TouchableOpacity onPress={onPress}>
         <Image
           style={ComponentStyle.imageStyle}
