+++ src/containers/shared/SplashPage/styles.js
