+++ src/containers/signedin/SettingsPage/UserPage/styles.js
@@ -1,0 +1,12 @@
+import { Dimensions } from "react-native";
+
+const SCREEN_WIDTH = Dimensions.get("window").width;
+const SCREEN_HEIGHT = Dimensions.get("window").height;
+
+export default (PageStyle = {
+  container: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center"
+  }
+});
